AccountLedgerScreen.kt fully linked finished testing remaining.

+++ app/src/main/java/com/vky342/openerp/data/Entities/Payment.kt
@@ -8,6 +8,6 @@ data class Payment(
     @PrimaryKey(autoGenerate = true)
     val paymentId : Int,
     val paymentDate : String,
-    val paymentAmount : Int,
+    val paymentAmount : Double,
     val ledgerId : Int
 )
