+++ app/src/main/java/com/vky342/openerp/data/Entities/Receipt.kt
@@ -9,6 +9,6 @@ data class Receipt(
     @PrimaryKey(autoGenerate = true)
     val receiptId : Int,
     val receiptDate : String,
-    val receiptAmount : Int,
+    val receiptAmount : Double,
     val ledgerId : String
 )
